First draft of a 'Chip sibling' class which we could use to drive module configs for PDK, lib, flow, etc setup
