+++ siliconcompiler/__init__.py
@@ -1,4 +1,8 @@
-from siliconcompiler.core import *
+from siliconcompiler.core.schema_consumer import *
+from siliconcompiler.core.error import *
+
+from siliconcompiler.core.chip import *
+from siliconcompiler.core.pdk import *
 #from siliconcompiler.floorplan import *
 #from siliconcompiler.leflib import *
 #from siliconcompiler.deflib import *
